Dashboard moved to page from unnecessary funcntion
