--- lib/clientUtils.ts
@@ -1,8 +1,0 @@
-"use client";
-
-export const redirectToForm = () => {
-  window.open(
-    "https://docs.google.com/forms/d/e/1FAIpQLSfVcdizYAyHxaOlbGcNlGf-NewHqq744HKP5lCC1Sx58wbtzQ/viewform",
-    "_blank"
-  );
-};
